Added some things for Asistencias and created the Create and Get Dtos for Asistencias

+++ GestionApi/Controllers/AlumnosController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using BCrypt.Net;
 using Microsoft.EntityFrameworkCore;
 using GestionApi.Data;
 using GestionApi.Models;
@@ -9,9 +10,11 @@ namespace GestionApi.Controllers{
 	[ApiController]
 	public class AlumnosController : ControllerBase{
 		private readonly AppDbContext _context;
+		private readonly IConfiguration _config;
 
-		public AlumnosController(AppDbContext context){
+		public AlumnosController(AppDbContext context, IConfiguration config){
 			_context = context; 
+			_config = config;
 		}
 
 		// GET: /api/Alumnos/{id}
@@ -43,9 +46,15 @@ namespace GestionApi.Controllers{
 		}
 		
 		[HttpDelete("{id}")]
-		public async Task<IActionResult> DeleteAlumno(int id){
-			var alumno = await _context.Alumnos.FindAsync(id);
+		public async Task<IActionResult> DeleteAlumno(int id, [FromBody] AlumnoDeleteDto request){
 
+			if (string.IsNullOrEmpty(request.Password) || !BCrypt.Net.BCrypt.Verify(request.Password, _config["AdminConfig:AdminHash"])){
+				return Unauthorized(new { message = "Acceso denegado, te faltan permisos."});
+			}
+			var alumno = await _context.Alumnos.FindAsync(id);
+			if (alumno == null){
+				return NotFound(new { message = "No se encontró ningún alumno con esa ID"});
+			}
 			_context.Alumnos.Remove(alumno);
 
 			await _context.SaveChangesAsync();
